// frontend/src/services/generated/models/article.ts
/**
 * Generated by orval v7.2.0 🍺
 * Do not edit manually.
 * DOCUMENTATION
 * OpenAPI spec version: 1.0.0
 */
import type { ArticleAuthor } from './articleAuthor';
import type { ArticleBlocksItem } from './articleBlocksItem';
import type { ArticleCategory } from './articleCategory';
import type { ArticleCover } from './articleCover';
import type { ArticleCreatedBy } from './articleCreatedBy';
import type { ArticleLocalizationsItem } from './articleLocalizationsItem';
import type { ArticleUpdatedBy } from './articleUpdatedBy';

export interface Article {
  author?: ArticleAuthor;
  blocks?: ArticleBlocksItem[];
  category?: ArticleCategory;
  /** @pattern ^\d*$ */
  comments?: string;
  cover?: ArticleCover;
  createdAt?: string;
  createdBy?: ArticleCreatedBy;
  description?: string;
  documentId?: string;
  id?: number;
  /** @pattern ^\d*$ */
  likes?: string;
  locale?: string;
  localizations?: ArticleLocalizationsItem[];
  publishedAt?: string;
  /** @pattern ^\d*$ */
  shares?: string;
  slug?: string;
  title?: string;
  updatedAt?: string;
  updatedBy?: ArticleUpdatedBy;
}